archive/python_scanner/requirements.txt
web3==6.11.0
requests==2.31.0
python-dotenv==1.0.0
pandas==2.0.3
numpy==1.24.3
asyncio==3.4.3
aiohttp==3.8.5
python-telegram-bot==20.5
ccxt==4.0.85
